Refactor feature metadata specs
